test(engine): extract network-dependent bare-clone test to integration build tag
TestEnsureBareClone_NewDir_ClonesFromLocal asserts that .fabrik/ is created
before the clone attempt, which requires a real network call to fail at the git
layer (not os.MkdirAll). The ENAMETOOLONG approach fails earlier, breaking this
assertion. Per EC-3: mark with //go:build integration so it is excluded from
the default test suite and documents the network dependency explicitly.
