+++ engine/worktree_extra_test.go
@@ -154,33 +154,6 @@ func TestEnsureBareClone_ExistingDir_FetchesOnly(t *testing.T) {
 	}
 }
 
-func TestEnsureBareClone_NewDir_ClonesFromLocal(t *testing.T) {
-	skipIfNoGit(t)
-
-	// Create a local "source" git repo to clone from via file:// URL
-	srcDir := initBareRepo(t)
-
-	tmpDir := t.TempDir()
-
-	// Override cloneURL construction by temporarily monkey-patching isn't possible in Go,
-	// but we can test the error path: clone of a non-existent github URL fails.
-	// Instead we test that the function creates the .fabrik directory and returns an error
-	// when git clone fails (no real network in CI).
-	_, err := ensureBareClone(tmpDir, "nonexistent-owner-xyz", "nonexistent-repo-xyz-abc", "", false)
-	// We expect an error because the github URL doesn't exist. The key check is that
-	// the function attempts the clone and returns a wrapped error.
-	if err == nil {
-		t.Log("clone unexpectedly succeeded (real network access?)")
-	}
-	// The .fabrik parent dir should have been created even on failure
-	fabrikDir := filepath.Join(tmpDir, ".fabrik")
-	if _, statErr := os.Stat(fabrikDir); os.IsNotExist(statErr) {
-		t.Error(".fabrik directory should be created before clone attempt")
-	}
-
-	// Suppress unused variable warning from srcDir
-	_ = srcDir
-}
 
 func TestSetCommitterIdentity_SetsWhenUnset(t *testing.T) {
 	skipIfNoGit(t)
